make Adapter implement class Adapter implements FilteredAdapter, change an item of .travis.yml
