+++ src/Adapter.php
@@ -6,17 +6,22 @@ use Casbin\Model\Model;
 use Casbin\Persist\Adapter as AdapterContract;
 use TechOne\Database\Manager;
 use Casbin\Persist\AdapterHelper;
+use Casbin\Persist\FilteredAdapter;
+use Casbin\Persist\Adapters\Filter;
+use Casbin\Exceptions\InvalidFilterTypeException;
 
 /**
  * DatabaseAdapter.
  *
  * @author techlee@qq.com
  */
-class Adapter implements AdapterContract
+class Adapter implements AdapterContract, FilteredAdapter
 {
     use AdapterHelper;
 
     protected $config;
+
+    protected $filtered;
 
     protected $connection;
 
@@ -25,8 +30,68 @@ class Adapter implements AdapterContract
     public function __construct(array $config)
     {
         $this->config = $config;
+        $this->filtered = false;
         $this->connection = (new Manager($config))->getConnection();
         $this->initTable();
+    }
+
+        /**
+     * Returns true if the loaded policy has been filtered.
+     *
+     * @return bool
+     */
+    public function isFiltered(): bool
+    {
+        return $this->filtered;
+    }
+
+    /**
+     * Loads only policy rules that match the filter from storage.
+     *
+     * @param Model $model
+     * @param mixed $filter
+     *
+     * @throws CasbinException
+     */
+    public function loadFilteredPolicy(Model $model, $filter): void
+    {
+        // if $filter is empty, load all policies
+        if (is_null($filter)) {
+            $this->loadPolicy($model);
+            return;
+        }
+        // validate $filter is a instance of Filter    
+        if (!$filter instanceof Filter) {
+            throw new InvalidFilterTypeException('invalid filter type');
+        }
+        $type = '';
+        $filter = (array) $filter;
+        // choose which ptype to use
+        foreach($filter as $i => $v) {
+            if (!empty($v)) {
+                array_unshift($filter[$i], $i);
+                $type = $i;
+                break;
+            }
+        }
+        $sql = 'SELECT ptype, v0, v1, v2, v3, v4, v5 FROM '.$this->casbinRuleTableName . ' WHERE ';
+        $items = ['ptype', 'v0', 'v1', 'v2', 'v3', 'v4', 'v5'];
+        $temp = [];
+        $values = [];
+        foreach($items as $i => $item) {
+            if (isset($filter[$type][$i]) && !empty($filter[$type][$i])) {
+                array_push($temp, $item . '=:' . $item);
+                $values[$item] = $filter[$type][$i];
+            }
+        }
+        $sql .= implode(' and ', $temp);
+        $rows = $this->connection->query($sql, $values);
+        foreach($rows as $row) {
+            $line = implode(', ', $row);
+            $this->loadPolicyLine($line, $model);
+        }
+        
+        $this->filtered = true;
     }
 
     public static function newAdapter(array $config)
